reorganize 2d scalar test
