+++ psydac/api/tests/test_2d_poisson.py
@@ -4,9 +4,9 @@ from mpi4py import MPI
 from sympy import pi, cos, sin, symbols
 from sympy.utilities.lambdify import implemented_function
 import pytest
+import os
 
 from sympde.calculus import grad, dot
-from sympde.calculus import laplace
 from sympde.topology import ScalarFunctionSpace
 from sympde.topology import element_of
 from sympde.topology import NormalVector
@@ -17,6 +17,7 @@ from sympde.expr     import Norm
 from sympde.expr     import find, EssentialBC
 
 from psydac.api.discretization import discretize
+from psydac.api.settings       import PSYDAC_BACKEND_GPYCCEL, PSYDAC_BACKEND_NUMBA
 
 x,y,z = symbols('x1, x2, x3')
 #==============================================================================
@@ -37,7 +38,7 @@ def get_boundaries(*args):
 
 #==============================================================================
 def run_poisson_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
-        ncells, degree, comm=None):
+        ncells, degree, backend=None, comm=None):
 
     assert isinstance(dir_zero_boundary   , (list, tuple))
     assert isinstance(dir_nonzero_boundary, (list, tuple))
@@ -92,11 +93,11 @@ def run_poisson_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
     Vh = discretize(V, domain_h, degree=degree)
 
     # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh])
+    equation_h = discretize(equation, domain_h, [Vh, Vh], backend=backend)
 
     # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh)
-    h1norm_h = discretize(h1norm, domain_h, Vh)
+    l2norm_h = discretize(l2norm, domain_h, Vh, backend=backend)
+    h1norm_h = discretize(h1norm, domain_h, Vh, backend=backend)
 
     #+++++++++++++++++++++++++++++++
     # 3. Solution
@@ -111,164 +112,11 @@ def run_poisson_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
 
     return l2_error, h1_error
 
-#==============================================================================
-def run_laplace_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
-        ncells, degree, comm=None):
-
-    assert isinstance(dir_zero_boundary   , (list, tuple))
-    assert isinstance(dir_nonzero_boundary, (list, tuple))
-
-    #+++++++++++++++++++++++++++++++
-    # 1. Abstract model
-    #+++++++++++++++++++++++++++++++
-    domain = Square()
-
-    B_dirichlet_0 = Union(*[domain.get_boundary(**kw) for kw in dir_zero_boundary])
-    B_dirichlet_i = Union(*[domain.get_boundary(**kw) for kw in dir_nonzero_boundary])
-    B_dirichlet   = Union(B_dirichlet_0, B_dirichlet_i)
-    B_neumann = domain.boundary.complement(B_dirichlet)
-
-    V  = ScalarFunctionSpace('V', domain)
-    u  = element_of(V, name='u')
-    v  = element_of(V, name='v')
-    nn = NormalVector('nn')
-
-    # Bilinear form a: V x V --> R
-    a = BilinearForm((u, v), integral(domain, dot(grad(u), grad(v)) + u * v))
-
-    # Linear form l: V --> R
-    l0 = LinearForm(v, integral(domain, f * v))
-    if B_neumann:
-        l1 = LinearForm(v, integral(B_neumann, v * dot(grad(solution), nn)))
-        l  = LinearForm(v, l0(v) + l1(v))
-    else:
-        l = l0
-
-    # Dirichlet boundary conditions
-    bc = []
-    if B_dirichlet_0:  bc += [EssentialBC(u,        0, B_dirichlet_0)]
-    if B_dirichlet_i:  bc += [EssentialBC(u, solution, B_dirichlet_i)]
-
-    # Variational model
-    equation = find(u, forall=v, lhs=a(u, v), rhs=l(v), bc=bc)
-
-    # Error norms
-    error  = u - solution
-    l2norm = Norm(error, domain, kind='l2')
-    h1norm = Norm(error, domain, kind='h1')
-
-    #+++++++++++++++++++++++++++++++
-    # 2. Discretization
-    #+++++++++++++++++++++++++++++++
-
-    # Create computational domain from topological domain
-    domain_h = discretize(domain, ncells=ncells, comm=comm)
-
-    # Discrete spaces
-    Vh = discretize(V, domain_h, degree=degree)
-
-    # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh])
-
-    # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh)
-    h1norm_h = discretize(h1norm, domain_h, Vh)
-
-    #+++++++++++++++++++++++++++++++
-    # 3. Solution
-    #+++++++++++++++++++++++++++++++
-
-    # Solve linear system
-    uh = equation_h.solve()
-
-    # Compute error norms
-    l2_error = l2norm_h.assemble(u=uh)
-    h1_error = h1norm_h.assemble(u=uh)
-
-    return l2_error, h1_error
-
-#==============================================================================
-def run_biharmonic_2d_dir(solution, f, dir_zero_boundary, ncells, degree, comm=None):
-
-    assert isinstance(dir_zero_boundary, (list, tuple))
-
-    #+++++++++++++++++++++++++++++++
-    # 1. Abstract model
-    #+++++++++++++++++++++++++++++++
-    domain = Square()
-
-    B_dirichlet_0 = Union(*[domain.get_boundary(**kw) for kw in dir_zero_boundary])
-    B_dirichlet_i = domain.boundary.complement(B_dirichlet_0)
-
-    V  = ScalarFunctionSpace('V', domain)
-    u  = element_of(V, name='u')
-    v  = element_of(V, name='v')
-    nn = NormalVector('nn')
-
-    # Bilinear form a: V x V --> R
-    a = BilinearForm((u, v), integral(domain, laplace(u) * laplace(v)))
-
-    # Linear form l: V --> R
-    l = LinearForm(v, integral(domain, f * v))
-
-    # Essential boundary conditions
-    dn = lambda a: dot(grad(a), nn)
-    bc = []
-    if B_dirichlet_0:
-        bc += [EssentialBC(   u , 0, B_dirichlet_0)]
-        bc += [EssentialBC(dn(u), 0, B_dirichlet_0)]
-    if B_dirichlet_i:
-        bc += [EssentialBC(   u ,    solution , B_dirichlet_i)]
-        bc += [EssentialBC(dn(u), dn(solution), B_dirichlet_i)]
-
-    # Variational model
-    equation = find(u, forall=v, lhs=a(u, v), rhs=l(v), bc=bc)
-
-    # Error norms
-    error  = u - solution
-    l2norm = Norm(error, domain, kind='l2')
-    h1norm = Norm(error, domain, kind='h1')
-    h2norm = Norm(error, domain, kind='h2')
-
-    #+++++++++++++++++++++++++++++++
-    # 2. Discretization
-    #+++++++++++++++++++++++++++++++
-
-    # Create computational domain from topological domain
-    domain_h = discretize(domain, ncells=ncells, comm=comm)
-
-    # Discrete spaces
-    Vh = discretize(V, domain_h, degree=degree)
-
-    # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh])
-
-    # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh)
-    h1norm_h = discretize(h1norm, domain_h, Vh)
-    h2norm_h = discretize(h2norm, domain_h, Vh)
-
-    #+++++++++++++++++++++++++++++++
-    # 3. Solution
-    #+++++++++++++++++++++++++++++++
-
-    # Solve linear system
-    uh = equation_h.solve()
-
-    # Compute error norms
-    l2_error = l2norm_h.assemble(u=uh)
-    h1_error = h1norm_h.assemble(u=uh)
-    h2_error = h2norm_h.assemble(u=uh)
-
-    return l2_error, h1_error, h2_error
 
 ###############################################################################
 #            SERIAL TESTS
 ###############################################################################
 
-#==============================================================================
-# 2D Poisson's equation
-#==============================================================================
 def test_poisson_2d_dir0_1234():
 
     solution = sin(pi*x)*sin(pi*y)
@@ -549,92 +397,78 @@ def test_poisson_2d_dir0_1234_user_function():
     assert( abs(l2_error - expected_l2_error) < 1.e-7)
     assert( abs(h1_error - expected_h1_error) < 1.e-7)
 
-#==============================================================================
-# 2D "Laplace-like" equation
-#==============================================================================
-def test_laplace_2d_neu0_1234():
+#------------------------------------------------------------------------------
+def test_poisson_2d_dir0_1234_gpyccel():
 
-    solution = cos(pi*x)*cos(pi*y)
-    f        = (2.*pi**2 + 1.)*solution
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
 
-    dir_zero_boundary    = get_boundaries()
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
     dir_nonzero_boundary = get_boundaries()
 
-    l2_error, h1_error = run_laplace_2d(solution, f, dir_zero_boundary,
-            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2])
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2], backend=PSYDAC_BACKEND_GPYCCEL)
 
-    expected_l2_error =  0.0002172846538950129
-    expected_h1_error =  0.012984852988125026
+    expected_l2_error =  0.00021808678604760232
+    expected_h1_error =  0.013023570720360362
+
+    assert( abs(l2_error - expected_l2_error) < 1.e-7)
+    assert( abs(h1_error - expected_h1_error) < 1.e-7)
+
+#------------------------------------------------------------------------------
+def test_poisson_2d_dir0_1234_numba():
+
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
+
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
+    dir_nonzero_boundary = get_boundaries()
+
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],backend=PSYDAC_BACKEND_NUMBA)
+
+    expected_l2_error =  0.00021808678604760232
+    expected_h1_error =  0.013023570720360362
 
     assert( abs(l2_error - expected_l2_error) < 1.e-7)
     assert( abs(h1_error - expected_h1_error) < 1.e-7)
 
-#==============================================================================
-# 2D biharmonic equation
-#==============================================================================
-def test_biharmonic_2d_dir0_1234():
+###############################################################################
+#            OpenMP TESTS
+###############################################################################
 
-    solution = sin(pi * x)**2 * sin(pi * y)**2
-    f        = laplace(laplace(solution))
+def test_poisson_openmp_2d_dir0_1234():
 
-    dir_zero_boundary = get_boundaries(1, 2, 3, 4)
+    x,y,z = symbols('x1, x2, x3')
 
-    l2_error, h1_error, h2_error = run_biharmonic_2d_dir(solution, f,
-            dir_zero_boundary, ncells=[2**3, 2**3], degree=[3, 3])
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
 
-    expected_l2_error = 0.00019981371108040476
-    expected_h1_error = 0.0063205179028178295
-    expected_h2_error = 0.2123929568623994
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
+    dir_nonzero_boundary = get_boundaries()
+
+    # Activate multi threading 
+    PSYDAC_BACKEND_GPYCCEL           = PSYDAC_BACKEND_GPYCCEL.copy()
+    PSYDAC_BACKEND_GPYCCEL['openmp'] = True
+    os.environ['OMP_NUM_THREADS']    = "2"
+
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],
+            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_GPYCCEL)
+
+    expected_l2_error =  0.00021808678604760232
+    expected_h1_error =  0.013023570720360362
 
     assert( abs(l2_error - expected_l2_error) < 1.e-7)
     assert( abs(h1_error - expected_h1_error) < 1.e-7)
-    assert( abs(h2_error - expected_h2_error) < 1.e-7)
 
-#------------------------------------------------------------------------------
-@pytest.mark.xfail
-def test_biharmonic_2d_dir0_123_diri_4():
-
-    solution = sin(pi * x)**2 * sin(0.5*pi * y)**2
-    f        = laplace(laplace(solution))
-
-    dir_zero_boundary = get_boundaries(1, 2, 3)
-
-    l2_error, h1_error, h2_error = run_biharmonic_2d_dir(solution, f,
-            dir_zero_boundary, ncells=[2**3, 2**3], degree=[3, 3])
-
-    print()
-    print(l2_error)
-    print(h1_error)
-    print(h2_error)
-    print()
-
-    assert False
-
-#------------------------------------------------------------------------------
-@pytest.mark.xfail
-def test_biharmonic_2d_dir0_13_diri_24():
-
-    solution = sin(3*pi/2 * x)**2 * sin(3*pi/2 * y)**2
-    f        = laplace(laplace(solution))
-
-    dir_zero_boundary = get_boundaries(1, 3)
-
-    l2_error, h1_error, h2_error = run_biharmonic_2d_dir(solution, f,
-            dir_zero_boundary, ncells=[2**3, 2**3], degree=[3, 3])
-
-    print()
-    print(l2_error)
-    print(h1_error)
-    print(h2_error)
-    print()
-
-    assert False
+    # Delete OMP_NUM_THREADS
+    del os.environ['OMP_NUM_THREADS']
 
 ###############################################################################
 #            PARALLEL TESTS
 ###############################################################################
 
-#==============================================================================
 @pytest.mark.parallel
 def test_poisson_2d_dir0_1234_parallel():
 
@@ -654,6 +488,75 @@ def test_poisson_2d_dir0_1234_parallel():
     assert( abs(l2_error - expected_l2_error) < 1.e-7)
     assert( abs(h1_error - expected_h1_error) < 1.e-7)
 
+#------------------------------------------------------------------------------
+@pytest.mark.parallel
+def test_poisson_2d_dir0_1234_parallel_pyccel():
+
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
+
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
+    dir_nonzero_boundary = get_boundaries()
+
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],
+            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_GPYCCEL)
+
+    expected_l2_error =  0.00021808678604760232
+    expected_h1_error =  0.013023570720360362
+
+    assert( abs(l2_error - expected_l2_error) < 1.e-7)
+    assert( abs(h1_error - expected_h1_error) < 1.e-7)
+
+#------------------------------------------------------------------------------
+@pytest.mark.parallel
+def test_poisson_2d_dir0_1234_parallel_numba():
+
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
+
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
+    dir_nonzero_boundary = get_boundaries()
+
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],
+            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_NUMBA)
+
+    expected_l2_error =  0.00021808678604760232
+    expected_h1_error =  0.013023570720360362
+
+    assert( abs(l2_error - expected_l2_error) < 1.e-7)
+    assert( abs(h1_error - expected_h1_error) < 1.e-7)
+
+#------------------------------------------------------------------------------
+@pytest.mark.parallel
+def test_poisson_openmp_2d_dir0_1234_parallel():
+
+    x,y,z = symbols('x1, x2, x3')
+
+    solution = sin(pi*x)*sin(pi*y)
+    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
+
+    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
+    dir_nonzero_boundary = get_boundaries()
+
+    # Activate multi threading 
+    PSYDAC_BACKEND_GPYCCEL           = PSYDAC_BACKEND_GPYCCEL.copy()
+    PSYDAC_BACKEND_GPYCCEL['openmp'] = True
+    os.environ['OMP_NUM_THREADS']    = "2"
+
+    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
+            dir_nonzero_boundary, ncells=[2**4, 2**4], degree=[2, 2],
+            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_GPYCCEL)
+
+    expected_l2_error =  2.6130834310749216e-05
+    expected_h1_error =  0.00320767625406208
+
+    assert( abs(l2_error - expected_l2_error) < 1.e-7)
+    assert( abs(h1_error - expected_h1_error) < 1.e-7)
+
+    # Delete OMP_NUM_THREADS
+    del os.environ['OMP_NUM_THREADS']
 
 #==============================================================================
 # CLEAN UP SYMPY NAMESPACE
