--- psydac/api/tests/test_api_2d_scalar_backends.py
@@ -1,362 +1,0 @@
-# -*- coding: UTF-8 -*-
-
-from mpi4py import MPI
-from sympy import pi, sin, symbols
-import pytest
-
-from sympde.calculus import grad, dot
-from sympde.calculus import laplace
-from sympde.topology import ScalarFunctionSpace
-from sympde.topology import element_of
-from sympde.topology import NormalVector
-from sympde.topology import Square
-from sympde.topology import Union
-from sympde.expr     import BilinearForm, LinearForm, integral
-from sympde.expr     import Norm
-from sympde.expr     import find, EssentialBC
-
-from psydac.api.discretization import discretize
-from psydac.api.settings import PSYDAC_BACKEND_GPYCCEL, PSYDAC_BACKEND_NUMBA
-
-x,y,z = symbols('x1, x2, x3')
-#==============================================================================
-def get_boundaries(*args):
-
-    if not args:
-        return ()
-    else:
-        assert all(1 <= a <= 4 for a in args)
-        assert len(set(args)) == len(args)
-
-    boundaries = {1: {'axis': 0, 'ext': -1},
-                  2: {'axis': 0, 'ext':  1},
-                  3: {'axis': 1, 'ext': -1},
-                  4: {'axis': 1, 'ext':  1}}
-
-    return tuple(boundaries[i] for i in args)
-
-#==============================================================================
-def run_poisson_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
-        ncells, degree, backend=None, comm=None):
-
-    assert isinstance(dir_zero_boundary   , (list, tuple))
-    assert isinstance(dir_nonzero_boundary, (list, tuple))
-
-    #+++++++++++++++++++++++++++++++
-    # 1. Abstract model
-    #+++++++++++++++++++++++++++++++
-    domain = Square()
-
-    B_dirichlet_0 = Union(*[domain.get_boundary(**kw) for kw in dir_zero_boundary])
-    B_dirichlet_i = Union(*[domain.get_boundary(**kw) for kw in dir_nonzero_boundary])
-    B_dirichlet   = Union(B_dirichlet_0, B_dirichlet_i)
-    B_neumann = domain.boundary.complement(B_dirichlet)
-
-    V  = ScalarFunctionSpace('V', domain)
-    u  = element_of(V, name='u')
-    v  = element_of(V, name='v')
-    nn = NormalVector('nn')
-
-    # Bilinear form a: V x V --> R
-    a = BilinearForm((u, v), integral(domain, dot(grad(u), grad(v))))
-
-    # Linear form l: V --> R
-    l0 = LinearForm(v, integral(domain, f * v))
-    if B_neumann:
-        l1 = LinearForm(v, integral(B_neumann, v * dot(grad(solution), nn)))
-        l  = LinearForm(v, l0(v) + l1(v))
-    else:
-        l = l0
-
-    # Dirichlet boundary conditions
-    bc = []
-    if B_dirichlet_0:  bc += [EssentialBC(u,        0, B_dirichlet_0)]
-    if B_dirichlet_i:  bc += [EssentialBC(u, solution, B_dirichlet_i)]
-
-    # Variational model
-    equation = find(u, forall=v, lhs=a(u, v), rhs=l(v), bc=bc)
-
-    # Error norms
-    error  = u - solution
-    l2norm = Norm(error, domain, kind='l2')
-    h1norm = Norm(error, domain, kind='h1')
-
-    #+++++++++++++++++++++++++++++++
-    # 2. Discretization
-    #+++++++++++++++++++++++++++++++
-
-    # Create computational domain from topological domain
-    domain_h = discretize(domain, ncells=ncells, comm=comm)
-
-    # Discrete spaces
-    Vh = discretize(V, domain_h, degree=degree)
-
-    # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh], backend=backend)
-
-    # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh, backend=backend)
-    h1norm_h = discretize(h1norm, domain_h, Vh, backend=backend)
-
-    #+++++++++++++++++++++++++++++++
-    # 3. Solution
-    #+++++++++++++++++++++++++++++++
-
-    # Solve linear system
-    uh = equation_h.solve()
-
-    # Compute error norms
-    l2_error = l2norm_h.assemble(u=uh)
-    h1_error = h1norm_h.assemble(u=uh)
-
-    return l2_error, h1_error
-
-#==============================================================================
-def run_laplace_2d(solution, f, dir_zero_boundary, dir_nonzero_boundary,
-        ncells, degree, comm=None):
-
-    assert isinstance(dir_zero_boundary   , (list, tuple))
-    assert isinstance(dir_nonzero_boundary, (list, tuple))
-
-    #+++++++++++++++++++++++++++++++
-    # 1. Abstract model
-    #+++++++++++++++++++++++++++++++
-    domain = Square()
-
-    B_dirichlet_0 = Union(*[domain.get_boundary(**kw) for kw in dir_zero_boundary])
-    B_dirichlet_i = Union(*[domain.get_boundary(**kw) for kw in dir_nonzero_boundary])
-    B_dirichlet   = Union(B_dirichlet_0, B_dirichlet_i)
-    B_neumann = domain.boundary.complement(B_dirichlet)
-
-    V  = ScalarFunctionSpace('V', domain)
-    u  = element_of(V, name='u')
-    v  = element_of(V, name='v')
-    nn = NormalVector('nn')
-
-    # Bilinear form a: V x V --> R
-    a = BilinearForm((u, v), integral(domain, dot(grad(u), grad(v)) + u * v))
-
-    # Linear form l: V --> R
-    l0 = LinearForm(v, integral(domain, f * v))
-    if B_neumann:
-        l1 = LinearForm(v, integral(B_neumann, v * dot(grad(solution), nn)))
-        l  = LinearForm(v, l0(v) + l1(v))
-    else:
-        l = l0
-
-    # Dirichlet boundary conditions
-    bc = []
-    if B_dirichlet_0:  bc += [EssentialBC(u,        0, B_dirichlet_0)]
-    if B_dirichlet_i:  bc += [EssentialBC(u, solution, B_dirichlet_i)]
-
-    # Variational model
-    equation = find(u, forall=v, lhs=a(u, v), rhs=l(v), bc=bc)
-
-    # Error norms
-    error  = u - solution
-    l2norm = Norm(error, domain, kind='l2')
-    h1norm = Norm(error, domain, kind='h1')
-
-    #+++++++++++++++++++++++++++++++
-    # 2. Discretization
-    #+++++++++++++++++++++++++++++++
-
-    # Create computational domain from topological domain
-    domain_h = discretize(domain, ncells=ncells, comm=comm)
-
-    # Discrete spaces
-    Vh = discretize(V, domain_h, degree=degree)
-
-    # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh])
-
-    # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh)
-    h1norm_h = discretize(h1norm, domain_h, Vh)
-
-    #+++++++++++++++++++++++++++++++
-    # 3. Solution
-    #+++++++++++++++++++++++++++++++
-
-    # Solve linear system
-    uh = equation_h.solve()
-
-    # Compute error norms
-    l2_error = l2norm_h.assemble(u=uh)
-    h1_error = h1norm_h.assemble(u=uh)
-
-    return l2_error, h1_error
-
-#==============================================================================
-def run_biharmonic_2d_dir(solution, f, dir_zero_boundary, ncells, degree, comm=None):
-
-    assert isinstance(dir_zero_boundary, (list, tuple))
-
-    #+++++++++++++++++++++++++++++++
-    # 1. Abstract model
-    #+++++++++++++++++++++++++++++++
-    domain = Square()
-
-    B_dirichlet_0 = Union(*[domain.get_boundary(**kw) for kw in dir_zero_boundary])
-    B_dirichlet_i = domain.boundary.complement(B_dirichlet_0)
-
-    V  = ScalarFunctionSpace('V', domain)
-    u  = element_of(V, name='u')
-    v  = element_of(V, name='v')
-    nn = NormalVector('nn')
-
-    # Bilinear form a: V x V --> R
-    a = BilinearForm((u, v), integral(domain, laplace(u) * laplace(v)))
-
-    # Linear form l: V --> R
-    l = LinearForm(v, integral(domain, f * v))
-
-    # Essential boundary conditions
-    dn = lambda a: dot(grad(a), nn)
-    bc = []
-    if B_dirichlet_0:
-        bc += [EssentialBC(   u , 0, B_dirichlet_0)]
-        bc += [EssentialBC(dn(u), 0, B_dirichlet_0)]
-    if B_dirichlet_i:
-        bc += [EssentialBC(   u ,    solution , B_dirichlet_i)]
-        bc += [EssentialBC(dn(u), dn(solution), B_dirichlet_i)]
-
-    # Variational model
-    equation = find(u, forall=v, lhs=a(u, v), rhs=l(v), bc=bc)
-
-    # Error norms
-    error  = u - solution
-    l2norm = Norm(error, domain, kind='l2')
-    h1norm = Norm(error, domain, kind='h1')
-    h2norm = Norm(error, domain, kind='h2')
-
-    #+++++++++++++++++++++++++++++++
-    # 2. Discretization
-    #+++++++++++++++++++++++++++++++
-
-    # Create computational domain from topological domain
-    domain_h = discretize(domain, ncells=ncells, comm=comm)
-
-    # Discrete spaces
-    Vh = discretize(V, domain_h, degree=degree)
-
-    # Discretize equation using Dirichlet bc
-    equation_h = discretize(equation, domain_h, [Vh, Vh])
-
-    # Discretize error norms
-    l2norm_h = discretize(l2norm, domain_h, Vh)
-    h1norm_h = discretize(h1norm, domain_h, Vh)
-    h2norm_h = discretize(h2norm, domain_h, Vh)
-
-    #+++++++++++++++++++++++++++++++
-    # 3. Solution
-    #+++++++++++++++++++++++++++++++
-
-    # Solve linear system
-    uh = equation_h.solve()
-
-    # Compute error norms
-    l2_error = l2norm_h.assemble(u=uh)
-    h1_error = h1norm_h.assemble(u=uh)
-    h2_error = h2norm_h.assemble(u=uh)
-
-    return l2_error, h1_error, h2_error
-
-###############################################################################
-#            SERIAL TESTS
-###############################################################################
-
-#==============================================================================
-# 2D Poisson's equation
-#==============================================================================
-def test_poisson_2d_dir0_1234_gpyccel():
-
-    solution = sin(pi*x)*sin(pi*y)
-    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
-
-    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
-    dir_nonzero_boundary = get_boundaries()
-
-    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
-            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2], backend=PSYDAC_BACKEND_GPYCCEL)
-
-    expected_l2_error =  0.00021808678604760232
-    expected_h1_error =  0.013023570720360362
-
-    assert( abs(l2_error - expected_l2_error) < 1.e-7)
-    assert( abs(h1_error - expected_h1_error) < 1.e-7)
-
-
-def test_poisson_2d_dir0_1234_numba():
-
-    solution = sin(pi*x)*sin(pi*y)
-    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
-
-    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
-    dir_nonzero_boundary = get_boundaries()
-
-    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
-            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2], backend=PSYDAC_BACKEND_NUMBA)
-
-    expected_l2_error =  0.00021808678604760232
-    expected_h1_error =  0.013023570720360362
-
-    assert( abs(l2_error - expected_l2_error) < 1.e-7)
-    assert( abs(h1_error - expected_h1_error) < 1.e-7)
-
-###############################################################################
-#            PARALLEL TESTS
-###############################################################################
-
-#==============================================================================
-
-@pytest.mark.parallel
-def test_poisson_2d_dir0_1234_parallel_pyccel():
-
-    solution = sin(pi*x)*sin(pi*y)
-    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
-
-    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
-    dir_nonzero_boundary = get_boundaries()
-
-    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
-            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],
-            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_GPYCCEL)
-
-    expected_l2_error =  0.00021808678604760232
-    expected_h1_error =  0.013023570720360362
-
-    assert( abs(l2_error - expected_l2_error) < 1.e-7)
-    assert( abs(h1_error - expected_h1_error) < 1.e-7)
-
-@pytest.mark.parallel
-def test_poisson_2d_dir0_1234_parallel_numba():
-
-    solution = sin(pi*x)*sin(pi*y)
-    f        = 2*pi**2*sin(pi*x)*sin(pi*y)
-
-    dir_zero_boundary    = get_boundaries(1, 2, 3, 4)
-    dir_nonzero_boundary = get_boundaries()
-
-    l2_error, h1_error = run_poisson_2d(solution, f, dir_zero_boundary,
-            dir_nonzero_boundary, ncells=[2**3, 2**3], degree=[2, 2],
-            comm=MPI.COMM_WORLD, backend=PSYDAC_BACKEND_NUMBA)
-
-    expected_l2_error =  0.00021808678604760232
-    expected_h1_error =  0.013023570720360362
-
-    assert( abs(l2_error - expected_l2_error) < 1.e-7)
-    assert( abs(h1_error - expected_h1_error) < 1.e-7)
-
-
-#==============================================================================
-# CLEAN UP SYMPY NAMESPACE
-#==============================================================================
-
-def teardown_module():
-    from sympy.core import cache
-    cache.clear_cache()
-
-def teardown_function():
-    from sympy.core import cache
-    cache.clear_cache()
